MCSC 354- Dev - Create Adj/Total Income Input Pages (#41)

+++ app/pages/annualallowance/taxyear/AdjustedIncomePage.scala
@@ -21,7 +21,7 @@ import pages.QuestionPage
 import play.api.libs.json.JsPath
 import play.api.mvc.Call
 
-case class AdjustedIncomePage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[Int] {
+case class AdjustedIncomePage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[BigInt] {
 
   override def path: JsPath = JsPath \ "aa" \ "years" \ period.toString \ "schemes" \ schemeIndex.toString \ toString
